refactor(web): step 6 — extract sessions CRUD + fork-anchor tracking to store/chat/session-list.ts
