+++ apps/web/src/store/chat/session-list.ts
@@ -1,0 +1,417 @@
+import { computed, ref, type Ref } from 'vue'
+import type { SessionSummary, UITurn } from '@/composables/api/useChat'
+import {
+  isSessionVisibleInSidebarMode,
+  sortByRecency,
+  type SidebarSessionMode,
+} from '../chat-list.utils'
+import { serverMessageId } from '../chat-list.normalize'
+import type { ChatMessage } from '../chat-list'
+
+// Sessions-list bookkeeping + fork-anchor tracking.
+//
+// These two concerns live in ONE module on purpose: `preserveSessionSummary`
+// (fork-anchor metadata preservation) is invoked inside every session-list
+// mutation, so splitting them would create a circular seam for no benefit.
+//
+// The factory owns the list state (reactive array + non-reactive O(1) map +
+// remembered/deleted bookkeeping) and every local mutation. Anything that
+// calls a transport (fetch/delete/rename) or orchestrates across other
+// clusters stays in the chat store.
+
+export interface SessionListDeps {
+  currentBotId: Ref<string | null>
+  sessionId: Ref<string | null>
+  // The active transcript, read (never written) by fork-anchor relocation.
+  messages: ChatMessage[]
+}
+
+export function createSessionList({ currentBotId, sessionId, messages }: SessionListDeps) {
+  const sessions = ref<SessionSummary[]>([])
+  // O(1) lookup keeps event handlers off the list scan that previously
+  // blocked the UI on bots with thousands of heartbeat sessions.
+  const sessionById = new Map<string, SessionSummary>()
+  const sessionLookupRevision = ref(0)
+  const rememberedSessions = ref<Record<string, SessionSummary>>({})
+  const deletedSessionIdsByBot = new Map<string, Set<string>>()
+  const sessionsCursor = ref<string | null>(null)
+  const hasMoreSessions = ref(false)
+  const loadingMoreSessions = ref(false)
+
+  const activeSession = computed(() => knownSessionSummary(sessionId.value ?? ''))
+  const knownSessions = computed<SessionSummary[]>(() => {
+    const byId = new Map<string, SessionSummary>()
+    for (const session of sessions.value) byId.set(session.id, session)
+    for (const session of Object.values(rememberedSessions.value)) {
+      if (!byId.has(session.id)) byId.set(session.id, session)
+    }
+    return [...byId.values()]
+  })
+
+  const activeChatReadOnly = computed(() => {
+    const session = activeSession.value
+    if (!session) return false
+    const type = session.type ?? 'chat'
+    if (type === 'heartbeat' || type === 'schedule' || type === 'subagent') return true
+    const ct = (session.channel_type ?? '').trim().toLowerCase()
+    if (ct && ct !== 'local') return true
+    return false
+  })
+
+  const activeChatCanFork = computed(() => activeSession.value?.type === 'chat')
+
+  function markSessionLookupChanged() {
+    sessionLookupRevision.value++
+  }
+
+  function trackSessionLookupRevision() {
+    return sessionLookupRevision.value
+  }
+
+  // --- Fork-anchor tracking -------------------------------------------------
+  // The anchor is stored as metadata.forked_from.fork_message_id on the
+  // SessionSummary; there is no dedicated state beyond the session records.
+
+  function forkSourceMetadata(session: SessionSummary | null | undefined): Record<string, unknown> | null {
+    const metadata = session?.metadata
+    if (!metadata || typeof metadata !== 'object') return null
+    const raw = (metadata as Record<string, unknown>).forked_from
+    return raw && typeof raw === 'object' ? raw as Record<string, unknown> : null
+  }
+
+  function forkSourceAnchor(session: SessionSummary | null | undefined): string {
+    return String(forkSourceMetadata(session)?.fork_message_id ?? '').trim()
+  }
+
+  function forkAnchorFromUITurns(turns: UITurn[], session?: SessionSummary | null): string {
+    const cutoff = Date.parse(session?.created_at ?? '')
+    const hasCutoff = Number.isFinite(cutoff)
+    for (let i = turns.length - 1; i >= 0; i--) {
+      const turn = turns[i]
+      if (turn.role !== 'assistant') continue
+      const id = String(turn.id ?? '').trim()
+      if (!id) continue
+      if (!hasCutoff) return id
+      const timestamp = Date.parse(turn.timestamp)
+      if (Number.isFinite(timestamp) && timestamp <= cutoff) return id
+    }
+    return ''
+  }
+
+  function withForkAnchorFromUITurns(session: SessionSummary, turns: UITurn[]): SessionSummary {
+    const anchor = forkAnchorFromUITurns(turns, session)
+    if (!anchor) return session
+    const fork = forkSourceMetadata(session)
+    if (!fork) return session
+    if (forkSourceAnchor(session) === anchor) return session
+
+    const metadata = session.metadata && typeof session.metadata === 'object' ? session.metadata : {}
+    return {
+      ...session,
+      metadata: {
+        ...metadata,
+        forked_from: {
+          ...fork,
+          fork_message_id: anchor,
+        },
+      },
+    }
+  }
+
+  function syncForkAnchorFromUITurns(targetSessionId: string | undefined, turns: UITurn[]) {
+    const sid = (targetSessionId ?? sessionId.value ?? '').trim()
+    if (!sid || turns.length === 0) return
+    const known = knownSessionSummary(sid)
+    if (!known || !forkSourceMetadata(known)) return
+    const anchored = withForkAnchorFromUITurns(known, turns)
+    if (anchored === known) return
+    upsertSession(anchored)
+    rememberSession(anchored)
+  }
+
+  function forkMessageCandidates(message: ChatMessage): string[] {
+    const out = [
+      message.serverId,
+      message.id,
+      message.role === 'system' ? undefined : message.externalMessageId,
+    ]
+      .map(value => value?.trim() ?? '')
+      .filter(Boolean)
+    return Array.from(new Set(out))
+  }
+
+  function messageMatchesForkAnchor(message: ChatMessage, anchor: string): boolean {
+    const target = anchor.trim()
+    return Boolean(target) && forkMessageCandidates(message).includes(target)
+  }
+
+  function latestInheritedAssistantBefore(index: number, session?: SessionSummary | null): string {
+    const cutoff = Date.parse(session?.created_at ?? '')
+    const hasCutoff = Number.isFinite(cutoff)
+    for (let i = index - 1; i >= 0; i--) {
+      const message = messages[i]
+      if (message?.role !== 'assistant') continue
+      if (hasCutoff) {
+        const timestamp = Date.parse(message.timestamp)
+        if (!Number.isFinite(timestamp) || timestamp > cutoff) continue
+      }
+      return serverMessageId(message) || message.id
+    }
+    return ''
+  }
+
+  function updateForkAnchorForReplacedMessage(targetSessionId: string, target: ChatMessage): (() => void) | null {
+    const sid = targetSessionId.trim()
+    if (!sid) return null
+    const known = knownSessionSummary(sid)
+    const fork = forkSourceMetadata(known)
+    const currentAnchor = String(fork?.fork_message_id ?? '').trim()
+    if (!known || !fork || !currentAnchor) return null
+
+    const targetIndex = messages.indexOf(target)
+    if (targetIndex < 0) return null
+    const replacedTailContainsAnchor = messages
+      .slice(targetIndex)
+      .some(message => messageMatchesForkAnchor(message, currentAnchor))
+    if (!replacedTailContainsAnchor) return null
+    const nextAnchor = latestInheritedAssistantBefore(targetIndex, known)
+    if (nextAnchor === currentAnchor) return null
+
+    const metadata = known.metadata && typeof known.metadata === 'object' ? known.metadata : {}
+    const nextFork = { ...fork }
+    if (nextAnchor) {
+      nextFork.fork_message_id = nextAnchor
+    } else {
+      delete nextFork.fork_message_id
+    }
+    const next = {
+      ...known,
+      metadata: {
+        ...metadata,
+        forked_from: nextFork,
+      },
+    }
+    replaceKnownSessionSummary(next)
+    rememberSession(next)
+    return () => {
+      replaceKnownSessionSummary(known)
+      rememberSession(known)
+    }
+  }
+
+  function preserveSessionSummary(incoming: SessionSummary, known?: SessionSummary | null): SessionSummary {
+    let next = incoming
+    if (known && !(next.title ?? '').trim() && (known.title ?? '').trim()) {
+      next = { ...next, title: known.title }
+    }
+
+    const knownFork = forkSourceMetadata(known)
+    if (!knownFork || !forkSourceAnchor(known) || forkSourceAnchor(next)) {
+      return next
+    }
+
+    const incomingFork = forkSourceMetadata(next)
+    const metadata = next.metadata && typeof next.metadata === 'object' ? next.metadata : {}
+    return {
+      ...next,
+      metadata: {
+        ...metadata,
+        forked_from: {
+          ...knownFork,
+          ...(incomingFork ?? {}),
+          fork_message_id: knownFork.fork_message_id,
+        },
+      },
+    }
+  }
+
+  // --- Session-list mutations ------------------------------------------------
+
+  function replaceSessions(items: SessionSummary[]): SessionSummary[] {
+    const currentDeleted = deletedSessionIdsByBot.get((currentBotId.value ?? '').trim())
+    // A racing list refresh can fetch a session before the backend's
+    // title-generation flow has persisted a title, while the client already
+    // holds one — the optimistic provisional title set in ensureActiveSession,
+    // which is local-only and never sent to the server. (Server-published
+    // titles don't have this problem: applyFallbackTitle and the LLM path
+    // both UpdateTitle before publishing, so the DB is current by the time any
+    // client sees the SSE.) An empty title in the snapshot means "server
+    // hasn't set one yet," not "title cleared," so preserve our non-empty title
+    // instead of letting the refresh erase it (which split the sidebar from
+    // the sticky tab title).
+    const merged = items
+      .filter(s => !currentDeleted?.has(s.id))
+      .map(s => preserveSessionSummary(s, sessionById.get(s.id) ?? rememberedSessions.value[s.id]))
+    sessions.value = merged
+    sessionById.clear()
+    for (const s of merged) sessionById.set(s.id, s)
+    markSessionLookupChanged()
+    return merged
+  }
+
+  function appendSessions(items: SessionSummary[]) {
+    if (items.length === 0) return
+    const currentDeleted = deletedSessionIdsByBot.get((currentBotId.value ?? '').trim())
+    const fresh = items
+      .filter(s => !sessionById.has(s.id) && !currentDeleted?.has(s.id))
+      .map(s => preserveSessionSummary(s, rememberedSessions.value[s.id]))
+    if (fresh.length === 0) return
+    sessions.value = [...sessions.value, ...fresh]
+    for (const s of fresh) sessionById.set(s.id, s)
+    markSessionLookupChanged()
+  }
+
+  function upsertSession(updated: SessionSummary) {
+    const currentDeleted = deletedSessionIdsByBot.get((currentBotId.value ?? '').trim())
+    if (currentDeleted?.has(updated.id)) return
+    const existing = sessionById.get(updated.id)
+    const remembered = rememberedSessions.value[updated.id]
+    const next = preserveSessionSummary(updated, existing ?? remembered)
+    if (existing) {
+      const rest = sessions.value.filter(session => session.id !== updated.id)
+      sessions.value = [next, ...rest]
+    } else {
+      sessions.value = [next, ...sessions.value]
+    }
+    sessionById.set(next.id, next)
+    markSessionLookupChanged()
+    if (remembered) rememberSession(next)
+  }
+
+  function replaceKnownSessionSummary(updated: SessionSummary) {
+    const currentDeleted = deletedSessionIdsByBot.get((currentBotId.value ?? '').trim())
+    if (currentDeleted?.has(updated.id)) return
+    if (sessionById.has(updated.id)) {
+      sessions.value = sessions.value.map(session => (session.id === updated.id ? updated : session))
+    }
+    sessionById.set(updated.id, updated)
+    markSessionLookupChanged()
+  }
+
+  function rememberSession(updated: SessionSummary) {
+    rememberedSessions.value = {
+      ...rememberedSessions.value,
+      [updated.id]: updated,
+    }
+    markSessionLookupChanged()
+  }
+
+  function forgetRememberedSession(id: string) {
+    if (!rememberedSessions.value[id]) return
+    const next = { ...rememberedSessions.value }
+    delete next[id]
+    rememberedSessions.value = next
+    markSessionLookupChanged()
+  }
+
+  function knownSessionSummary(targetSessionId: string): SessionSummary | null {
+    const sid = targetSessionId.trim()
+    if (!sid) return null
+    // `sessionById` is intentionally a non-reactive O(1) lookup. Tie computed
+    // consumers such as activeChatTarget to a tiny revision so an early null
+    // read is invalidated when the session list/hydration later fills the map.
+    trackSessionLookupRevision()
+    return sessionById.get(sid) ?? rememberedSessions.value[sid] ?? null
+  }
+
+  function isRecentsSession(session: SessionSummary): boolean {
+    const type = (session.type ?? 'chat').trim()
+    return type === 'chat' || type === 'discuss' || type === 'acp_agent'
+  }
+
+  // patchSessionInList applies a partial update to one session in BOTH the
+  // reactive `sessions` array (reassigned so the sidebar virtualizer and any
+  // `sessions`-derived computed re-run) and the `sessionById` lookup map. SSE
+  // title/touch handlers must route through this: mutating the map's stored
+  // object in place (`target.title = ...`) writes the raw object but never
+  // triggers `sessions.value`, so the UI stays stale until a full REST refresh
+  // (the Cmd+R symptom).
+  function patchSessionInList(id: string, patch: Partial<SessionSummary>) {
+    const existing = sessionById.get(id)
+    if (!existing) return
+    const currentDeleted = deletedSessionIdsByBot.get((existing.bot_id ?? currentBotId.value ?? '').trim())
+    if (currentDeleted?.has(id)) return
+    const next = { ...existing, ...patch }
+    sessionById.set(id, next)
+    sessions.value = sessions.value.map(session => (session.id === id ? next : session))
+    markSessionLookupChanged()
+  }
+
+  function removeSessionFromList(id: string) {
+    if (!sessionById.has(id) && !rememberedSessions.value[id]) return
+    sessions.value = sessions.value.filter(session => session.id !== id)
+    sessionById.delete(id)
+    markSessionLookupChanged()
+    forgetRememberedSession(id)
+  }
+
+  function touchSessionInList(targetSessionId: string, timestamp?: string) {
+    const target = sessionById.get(targetSessionId)
+    if (!target) return
+    if (timestamp && (!target.updated_at || timestamp > target.updated_at)) {
+      patchSessionInList(targetSessionId, { updated_at: timestamp })
+    }
+  }
+
+  function fallbackSessionAfterDelete(mode: SidebarSessionMode): SessionSummary | null {
+    const visibleSessions = sessions.value.filter(session => isSessionVisibleInSidebarMode(session, mode))
+    return sortByRecency(visibleSessions)[0] ?? null
+  }
+
+  function markSessionDeleted(botId: string, targetSessionId: string) {
+    const bid = botId.trim()
+    const sid = targetSessionId.trim()
+    if (!bid || !sid) return
+    const deletedIds = deletedSessionIdsByBot.get(bid) ?? new Set<string>()
+    deletedIds.add(sid)
+    deletedSessionIdsByBot.set(bid, deletedIds)
+  }
+
+  function clearDeletedSessionIds() {
+    deletedSessionIdsByBot.clear()
+  }
+
+  function clearRememberedSessions() {
+    rememberedSessions.value = {}
+  }
+
+  return {
+    // state
+    sessions,
+    sessionById,
+    rememberedSessions,
+    sessionsCursor,
+    hasMoreSessions,
+    loadingMoreSessions,
+    // computeds
+    activeSession,
+    knownSessions,
+    activeChatReadOnly,
+    activeChatCanFork,
+    // fork anchor
+    forkSourceMetadata,
+    forkSourceAnchor,
+    forkAnchorFromUITurns,
+    withForkAnchorFromUITurns,
+    syncForkAnchorFromUITurns,
+    messageMatchesForkAnchor,
+    updateForkAnchorForReplacedMessage,
+    preserveSessionSummary,
+    // mutations
+    replaceSessions,
+    appendSessions,
+    upsertSession,
+    replaceKnownSessionSummary,
+    rememberSession,
+    forgetRememberedSession,
+    knownSessionSummary,
+    isRecentsSession,
+    patchSessionInList,
+    removeSessionFromList,
+    touchSessionInList,
+    fallbackSessionAfterDelete,
+    markSessionDeleted,
+    clearDeletedSessionIds,
+    clearRememberedSessions,
+  }
+}
